Update CurTime to double

+++ source/gmsv_turbostroi.h
@@ -6,6 +6,6 @@
 
 extern std::string g_LibraryFileName;
 extern bool g_ForceThreadsFinished;
-extern std::atomic<float> g_CurrentTime;
+extern std::atomic<double> g_CurrentTime;
 extern unsigned int g_ThreadTickrate;
 extern std::unordered_map<std::string, std::string> g_LoadedFilesCache;
